Moved duplicated code into a new npm library 'kaw-users-utils'
