+++ users/authservice/package.json
@@ -24,6 +24,7 @@
     "body-parser": "^1.20.2",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.2",
+    "kaw-users-utils": "^1.0.5",
     "mongoose": "^8.0.4"
   },
   "devDependencies": {
