+++ users/authservice/src/models/auth-model.ts
@@ -1,26 +1,7 @@
 import mongoose from 'mongoose';
+import { userJson } from 'kaw-users-utils';
 
-const userSchema = new mongoose.Schema({
-  username: {
-    type: String,
-    required: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-  createdAt: {
-    type: Date,
-    default: Date.now,
-  },
-  history: {
-    passedQuestions: { type: Number, default: 0 },
-    failedQuestions: { type: Number, default: 0 },
-    gamesPlayed: { type: Number, default: 0 },
-    timePlayed: { type: Number, default: 0 },
-    points: { type: Number, default: 0 },
-  },
-});
+const userSchema = new mongoose.Schema(userJson);
 
 const User = mongoose.model('User', userSchema);
