+++ users/userservice/package.json
@@ -22,8 +22,9 @@
   "dependencies": {
     "bcrypt": "^5.1.1",
     "body-parser": "^1.20.2",
-    "express": "^4.18.2",
+    "express": "^4.19.1",
     "jsonwebtoken": "^9.0.2",
+    "kaw-users-utils": "^1.0.0",
     "mongoose": "^8.0.4"
   },
   "devDependencies": {
