+++ users/userservice/src/models/user-model.ts
@@ -1,27 +1,8 @@
 import mongoose from 'mongoose';
 import { IUser } from '../types';
+import { userJson } from 'kaw-users-utils';
 
-const userSchema = new mongoose.Schema<IUser>({
-  username: {
-    type: String,
-    required: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-  createdAt: {
-    type: Date,
-    default: Date.now,
-  },
-  history: {
-    passedQuestions: { type: Number, default: 0 },
-    failedQuestions: { type: Number, default: 0 },
-    gamesPlayed: { type: Number, default: 0 },
-    timePlayed: { type: Number, default: 0 },
-    points: { type: Number, default: 0 },
-  },
-});
+const userSchema = new mongoose.Schema<IUser>(userJson);
 
 const User = mongoose.model<IUser>('User', userSchema);
